requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sentence-transformers==2.2.2
faiss-cpu==1.13.1
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.2.3
tqdm==4.65.0
huggingface-hub==0.13.3
tqdm==4.65.0
tqdm==4.65.0
python-multipart==0.0.6
pytest==7.4.0
